IntranetAPI/requirements.txt
requests
tiktoken
python-dotenv
openai==0.28.0
qdrant-client==1.16.0
Flask
flask-cors
flask-limiter
uvicorn
asgiref
beautifulsoup4
pdfplumber==0.11.7
tzdata
playwright
python-docx